+++ apps/web/src/components/resume/HardSkills.tsx
