+++ apps/web/src/components/resume/SoftSkills.tsx
